Moco_v2 renamed to MoCo
